refactor: move exceptions into a single file
Since the most of the exceptions are for internal
use, the change must not be breaking for the end
users. Exceptions which are designed to be used
by the end users (`CacheBackendInteractionError`,
`LockedError`) are still available under the same
location.
