+++ cashews/__init__.py
@@ -1,6 +1,6 @@
 from ._cache_condition import NOT_NONE  # noqa
-from .backends.interface import CacheBackendInteractionError, LockedError  # noqa
 from .decorators import CacheDetect, CircuitBreakerOpen, RateLimitError, context_cache_detect, fast_condition  # noqa
+from .exceptions import CacheBackendInteractionError, LockedError  # noqa
 from .formatter import default_formatter, get_template_and_func_for, get_template_for_key  # noqa
 from .helpers import add_prefix, all_keys_lower, memory_limit  # noqa
 from .key import noself  # noqa
